+++ apps/web/src/app/(needLogin)/travel/_apis/planPostApi.ts
@@ -1,5 +1,4 @@
 // plan post fetch 요청 로직
-
 import { TravelPlanType, TravelRecordType } from '@/types/TravelRegister.types';
 
 export const postPlan = async (formdata: TravelPlanType) => {
@@ -12,7 +11,7 @@ export const postPlan = async (formdata: TravelPlanType) => {
   })
     .then(async (res) => {
       if (res.status === 200) {
-        alert('post 요청 성공');
+        localStorage.removeItem('formPlanAtom');
         return res;
       } else if (res.status === 403) {
         return res.json();
@@ -34,8 +33,8 @@ export const postRecord = async (formdata: TravelRecordType) => {
   })
     .then((res) => {
       if (res.status === 200) {
-        alert('post 요청 성공');
-        // localStorage.removeItem('formRecordAtom'); // formAtom 데이터 삭제할 때 사용
+        // alert('post 요청 성공');
+        localStorage.removeItem('formRecordAtom'); // formAtom 데이터 삭제할 때 사용
       } else if (res.status === 403) {
         return res.json();
       }
