+++ apps/web/src/app/(needLogin)/travel/_components/NavigateButton/NavigateButton.tsx
@@ -1,28 +1,10 @@
 'use client';
-import { useAtomValue } from 'jotai';
 import Link from 'next/link';
-import { useRouter, useSearchParams } from 'next/navigation';
 
-import { registerStateAtom } from '@/store/travelState.atom';
-
-import { useDateSelection } from './use-data-selection';
+import { useTravelForm } from '../../_hooks/useTravelRegister';
 
 const NavigateButton = ({ handleSubmit, onSubmit }) => {
-  const params = useSearchParams();
-  const search = Number(params.get('stepId'));
-  const registerState = useAtomValue(registerStateAtom);
-  const router = useRouter();
-  const { isDateSelected } = useDateSelection();
-
-  const handleNextButtonClick = (e) => {
-    e.preventDefault();
-    if (isDateSelected) {
-      handleSubmit(onSubmit)();
-      router.push(`/travel/${registerState}?stepId=${search + 1}`);
-    } else {
-      alert('날짜를 선택해주세요');
-    }
-  };
+  const { handleNextButtonClick, registerState, search } = useTravelForm();
 
   return (
     <>
@@ -44,7 +26,7 @@ const NavigateButton = ({ handleSubmit, onSubmit }) => {
         <span
           className='material-icons-outlined '
           style={{ fontSize: '50px', color: '#ECF1F7' }}
-          onClick={handleNextButtonClick}
+          onClick={handleNextButtonClick(handleSubmit, onSubmit)}
         >
           arrow_forward_ios
         </span>
