+++ apps/web/src/app/(needLogin)/travel/_components/Step3What/Step3What.tsx
@@ -1,15 +1,10 @@
 'use client';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useAtom, useSetAtom } from 'jotai';
+// import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
 
-import { formModePlanAtom, formModeRecordAtom } from '@/store';
-import { getStepBarAtom } from '@/store/stepNavbar.atom';
-import { registerStateAtom } from '@/store/travelState.atom';
-
 import { travelTag } from '../../../../../constants/travelStep3Tag.constants';
-import { postPlan, postRecord } from '../../_apis/planPostApi';
+import { useTravelForm } from '../../_hooks/useTravelRegister';
 import { CATEGORY_SCHEMA } from '../../_schema/travel.schema';
 import type { IStep3Props } from '../../Travel.type';
 import NavigateButton from '../NavigateButton/NavigateButton';
@@ -18,11 +13,8 @@ import Step3Tag from './components/Step3Tag/Step3Tag';
 
 const Step3What: React.FC<IStep3Props> = ({ config }) => {
   // const queryClient = useQueryClient();
-  const [registerState] = useAtom(registerStateAtom);
-  const [planAtom, setPlanAtom] = useAtom(formModePlanAtom);
-  const [recordAtom, setRecordAtom] = useAtom(formModeRecordAtom);
 
-  const setStepbarAtom = useSetAtom(getStepBarAtom(3));
+  const { registerState, step3onSubmit } = useTravelForm();
   // const mutation = useMutation(postPlan, {
   //   onSuccess: () => {
   //     queryClient.invalidateQueries();
@@ -32,35 +24,9 @@ const Step3What: React.FC<IStep3Props> = ({ config }) => {
     resolver: yupResolver(CATEGORY_SCHEMA),
   });
 
-  const onSubmit = async (data: {
-    tag0: string;
-    tag1: string;
-    tag2: string;
-    tag3: string;
-    theme: string;
-  }) => {
-    setStepbarAtom(true);
-    registerState == 'record'
-      ? setRecordAtom((prev) => ({
-          ...prev,
-          theme: data.theme.toString(),
-          travelHashTags: Array(4)
-            .fill(null)
-            .map((_, index) => ({
-              id: Math.floor(Math.random() * 100000),
-              hashTag: { name: data[`tag${index}`].slice(1) },
-            })),
-        }))
-      : setPlanAtom((prev) => ({
-          ...prev,
-          theme: data.theme.toString(),
-        }));
-    registerState == 'plan' ? postPlan(planAtom) : postRecord(recordAtom);
-  };
-
   return (
     <>
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <form onSubmit={handleSubmit(step3onSubmit)}>
         <div className='mt-[57px] flex h-[600px] items-center justify-center '>
           <div className='bg-primaryBlue-100 absolute flex h-[600px] w-full max-w-[969px] flex-col items-center justify-center'>
             <div>
@@ -90,7 +56,10 @@ const Step3What: React.FC<IStep3Props> = ({ config }) => {
               )}
             </div>
           </div>
-          <NavigateButton handleSubmit={handleSubmit} onSubmit={onSubmit} />
+          <NavigateButton
+            handleSubmit={handleSubmit}
+            onSubmit={step3onSubmit}
+          />
         </div>
       </form>
     </>
