+++ apps/web/src/app/(needLogin)/travel/_components/Step4How/Step4How.tsx
@@ -1,11 +1,7 @@
 'use client';
 
-import { useAtom } from 'jotai';
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
-
-import { formModePlanAtom, formModeRecordAtom } from '@/store';
-import { registerStateAtom } from '@/store/travelState.atom';
 
 import { IStep4Props } from '../../Travel.type';
 import NavigateButton from '../NavigateButton/NavigateButton';
@@ -14,24 +10,15 @@ import Step4Dates from './components/Step4Dates/Step4Dates';
 import { dates } from './mocks';
 
 const Step4How: React.FC<IStep4Props> = ({ config }) => {
-  const [registerState] = useAtom(registerStateAtom);
-  const [planAtom] = useAtom(formModePlanAtom);
-  const [recordAtom] = useAtom(formModeRecordAtom);
   const { handleSubmit, control } = useForm();
   const onSubmit = (data) => console.log(data);
 
   const [currentPage, setCurrentPage] = useState(0);
   const [itemsPerPage] = useState(4);
 
-  registerState == 'plan'
-    ? console.log('plan data', planAtom)
-    : console.log('record data', recordAtom);
-
   return (
     <>
-      <form
-      // onSubmit={handleSubmit(onSubmit)}
-      >
+      <form>
         <div>
           <div className='mt-[57px] flex h-[600px] items-center justify-center '>
             <div className='bg-primaryBlue-100 absolute flex h-[600px] w-full max-w-[969px] justify-center border'>
